Tests cleanup
* Removed duplicate try/catches as found
